moving error handler to separate file
