+++ lambda/custom/handlers/helperIntents.js
@@ -1,4 +1,3 @@
-const { ri } = require('@jargon/alexa-skill-sdk');
 const { customLog } = require('../helperFunctions');
 
 const SessionEndedRequestHandler = {
@@ -18,23 +17,6 @@ const SessionEndedRequestHandler = {
 	},
 };
 
-const ErrorHandler = {
-	canHandle() {
-		return true;
-	},
-	async handle(handlerInput, error) {
-		console.log(`Error handled: ${ error.message }`);
-		customLog({ errorMessage: error.message });
-		const speechText = ri('ERRORS');
-
-		return handlerInput.jrb
-			.speak(speechText)
-			.reprompt(speechText)
-			.getResponse();
-	},
-};
-
 module.exports = {
 	SessionEndedRequestHandler,
-	ErrorHandler,
 };
